Interlaced collectors for read_interlaced_*

Interlaced collector extend `readr` collector types (e.g. `col_double()`) to allow column-level missing value specifications.

Usage

icol_logical(na)

icol_integer(na)

icol_double(na)

icol_character(na)

icol_factor(na, levels = NULL, ordered = FALSE)

icol_date(na, format = "")

icol_time(na, format = "")

icol_datetime(na, format = "")

icol_number(na)

Arguments

na

Character vector of strings to interpret as column-level missing values

levels

Character vector of the allowed levels. When levels = NULL (the default), levels are discovered from the unique values of x, in the order in which they appear in x.

ordered

Is it an ordered factor?

format

A format specification, as described below. If set to "", date times are parsed as ISO8601, dates and times used the date and time formats specified in the locale().

Unlike strptime(), the format specification must match the complete string.