Test if a value is missing and lacks a missing reason

When a value is missing both a value and a missing reason, it is considered "empty". is.empty() checks for these type of values. Regular NA values (with no missing reasons) are also considered "empty".

Usage

is.empty(x)

Arguments

x

a vector

Value

a logical vector the same length as x, containing TRUE for all empty elements, and FALSE otherwise.