Missing reason collectors

Missing reason collectors are used in extended column specifications to specify the type of a column's missing reason channel.

Usage

na_col_default()

na_col_none()

na_col_integer(...)

na_col_factor(...)

na_col_cfactor(...)

Arguments

...

values to interpret as missing values. In the case of na_col_cfactor(), arguments must be named.

Value

a new missing reason collector object

Details

na_col_default() is used to signal that the missing reason type should inherit the specification provided in the na = argument of the calling read_interlaced_*() function

See also