Value collectors

Value collectors are used in extended column specifications to specify the value type of a column. They are think wrappers around readr's col_*() collector types.

Usage

v_col_guess()

v_col_cfactor(codes, ordered = FALSE)

v_col_character()

v_col_date(format = "")

v_col_datetime(format = "")

v_col_double()

v_col_factor(levels = NULL, ordered = FALSE)

v_col_integer()

v_col_big_integer()

v_col_logical()

v_col_number()

v_col_skip()

v_col_time(format = "")

Arguments

codes

A named vector of unique codes that declares the mapping of labels to codes.

ordered

Is it an ordered factor?

format

A format specification, as described in readr::col_datetime()

levels

Character vector of the allowed levels. When levels = NULL (the default), levels are discovered from the unique values of x, in the order in which they appear in x.

Value

a new value collector object

Details

In addition to all of the column types supported by readr, interlacer additionally can load cfactor() types via v_col_cfactor()

See also